(feature) modificacion datos, en montos, se modifica servicio de crear ventas

+++ src/main/java/com/geoprom/cl/api/backend/models/DTOs/ProductoDTO.java
@@ -7,6 +7,8 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ProductoDTO {
@@ -19,8 +21,8 @@ public class ProductoDTO {
     private String nombre;
     private String rut_empresa;
     private String sku;
-    private Integer precio_compra;
-    private Integer precio_venta;
+    private BigDecimal precio_compra;
+    private BigDecimal precio_venta;
     private Integer stock;
     private String urlImg;
     private Integer estado;
