+++ src/main/java/com/geoprom/cl/api/backend/models/Request/Productos/UpdateProductoRequest.java
@@ -4,6 +4,8 @@ package com.geoprom.cl.api.backend.models.Request.Productos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class UpdateProductoRequest {
@@ -12,8 +14,8 @@ public class UpdateProductoRequest {
     private String nombre;
     private String rut_empresa;
     private String sku;
-    private Integer precio_compra;
-    private Integer precio_venta;
+    private BigDecimal precio_compra;
+    private BigDecimal precio_venta;
     private Integer stock;
     private String urlImg;
     private Integer estado;
